refactor: rename tableComponentModule into InventorableComponentModule
