+++ apps/gauzy/src/app/pages/inventory/components/inventory-table-components/enabled-row.component.ts
