+++ apps/gauzy/src/app/pages/inventory/components/inventory-table-components/icon-row.component.ts
