+++ apps/gauzy/src/app/pages/inventory/components/inventory-table-components/item-img-tags-row.component.ts
