+++ apps/gauzy/src/app/pages/inventory/components/inventory-table-components/name-with-description/name-with-description.component.ts
