+++ apps/gauzy/src/app/pages/inventory/components/inventory-table-components/no-image/no-image.component.html
