+++ apps/gauzy/src/app/pages/inventory/components/inventory-table-components/selected-row.component.ts
