+++ apps/gauzy/src/app/pages/inventory/components/manage-merchants/merchant.module.ts
@@ -31,7 +31,7 @@ import { TagsColorInputModule } from '../../../../@shared/tags/tags-color-input/
 import { GauzyButtonActionModule } from './../../../../@shared/gauzy-button-action/gauzy-button-action.module';
 import { CardGridModule } from './../../../../@shared/card-grid/card-grid.module';
 import { NgxPermissionsModule } from 'ngx-permissions';
-import { TableComponentsModule } from '../table-components';
+import { InventoryTableComponentsModule } from '../inventory-table-components';
 
 const NB_MODULES = [
 	NbCardModule,
@@ -71,7 +71,7 @@ const NB_MODULES = [
         TagsColorInputModule,
         GauzyButtonActionModule,
         CardGridModule,
-        TableComponentsModule,
+        InventoryTableComponentsModule,
         NgxPermissionsModule.forChild()
     ],
     providers: [MerchantService]
