+++ apps/gauzy/src/app/pages/inventory/components/manage-product-categories/product-categories.component.ts
@@ -12,7 +12,7 @@ import {
 	ComponentLayoutStyleEnum
 } from '@gauzy/contracts';
 import { distinctUntilChange } from '@gauzy/common-angular';
-import { ImageRowComponent } from './../table-components';
+import { ImageRowComponent } from '../inventory-table-components';
 import { ProductCategoryMutationComponent } from '../../../../@shared/product-mutation';
 import { DeleteConfirmationComponent } from '../../../../@shared/user/forms';
 import { API_PREFIX, ComponentEnum } from '../../../../@core/constants';
