+++ apps/gauzy/src/app/pages/inventory/inventory.module.ts
@@ -49,7 +49,7 @@ import { MerchantModule } from './components/manage-merchants/merchant.module';
 import { ProductTypesModule } from './components/manage-product-types/product-types.module';
 import { ProductCategoriesModule } from './components/manage-product-categories/product-categories.module';
 import { WarehousesModule } from './components/manage-warehouses/warehouses.module';
-import { TableComponentsModule } from './components/table-components/table-components.module';
+import { InventoryTableComponentsModule } from './components/inventory-table-components/inventory-table-components.module';
 import { SelectAssetModule } from '../../@shared/select-asset-modal/select-asset.module';
 import { ImageAssetModule } from '../../@shared/image-asset/image-asset.module';
 import { PaginationModule } from '../../@shared/pagination/pagination.module';
@@ -101,7 +101,7 @@ const NB_MODULES = [
 		ReactiveFormsModule,
 		SharedModule,
 		SelectAssetModule,
-		TableComponentsModule,
+		InventoryTableComponentsModule,
 		TagsColorInputModule,
 		ThemeModule,
 		TranslateModule,
